Rf/cleanup (#22)
* Refile CodeViewer component into /student directory

* Cleanup some code organization

* Link standard comments to comments, primitive implementation

* Delete Comment component. Only use EditableComment

* Never force jsx boolean value

* Updating some comment styling

* Code alignment

* More cleanup

* Clickable rubric categories

* Add comment loading/save icon

* Fix stacking a bit

* Consolidate code utils

notably the highlighting code and comment size estimating.

The decision here was to keep CodeViewer / CodeGrader components
separate, since they may diverge more substantially in future
versions. However, it is annoying to update style and logic in one
and not have it reflect in the other. This utility class can help
solve that for now.